cprogram

+++ average.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 #include <stdlib.h>
 int main(void){
-    int num1, num2, num3, average;
+    float num1, num2, num3, average;
     printf("Enter 3 numbers : ");
-    scanf("%d %d %d", &num1 , &num2, &num3);
+    scanf("%f %f %f", &num1 , &num2, &num3);
     average = (num1 + num2 +num3)/3;
-    printf("average is %d\n",average);
+    printf("average is %f\n",average);
   return 0;
     
 }
